requirements.txt
tinkoff-investments
selenium
python-dotenv
ipdb